requirements.txt
torch>=2.0.0
transformers>=4.35.0
gradio>=4.0.0
Pillow>=9.0.0
accelerate>=0.20.0
